Move metadata.go at models.go
